changed another datatype in db

+++ db/migrate/20180404095324_create_results.rb
@@ -4,7 +4,7 @@ class CreateResults < ActiveRecord::Migration[5.2]
       t.references :athlete, foreign_key: true
       t.float :score, default: 0.0
       t.float :rating_collection, array: true, default: []
-      t.integer :number_of_votes, default: 0
+      t.float :number_of_votes, default: 0
 
       t.timestamps
     end
